+++ aries-site/src/examples/components/pageheader/index.js
@@ -2,5 +2,6 @@ export * from './ChildPageHeaderExample';
 export * from './PageHeaderActions';
 export * from './PageHeaderAnatomy';
 export * from './PageHeaderIntroExample';
+export * from './PageHeaderResponsiveActions';
 export * from './PageHeaderSubtitle';
 export * from './TopLevelPageHeaderExample';
